requirements.txt
Flask
barcodenumber
flask_sqlalchemy
flask_session
hashlib
google-genai
python-dotenv
